# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QuantLab Signal Viewer API
  version: 0.9.0
  description: >
    Typed API boundary for the QuantLab demo. This document is the authored
    single source of truth (Constitution V): the FastAPI app is validated
    against it by a contract test, and the TypeScript client is generated from
    it with openapi-typescript. All computation happens in the backend.

    Supersedes specs/005-signal-research-workbench/contracts/openapi.yaml as
    the authored home, carrying every existing path forward and adding dataset
    provenance for feature 006-warehouse-experiments. The same paths serve
    either the synthetic demo dataset or real ingested history; handlers do not
    branch on which, and every run records which one produced it. Running a
    model remains the one state-changing operation, and it writes only to
    experiment storage -- never to a curated or materialised signal set.

    0.6.0 adds authentication (feature 007). When auth is enabled
    (QUANTLAB_AUTH, default on), every operation except `getHealth`, `login`
    and the bootstrap-gated `registerUser` requires the `quantlab_session`
    cookie issued by `login`; unauthenticated calls answer 401. Experiment
    runs are scoped to their owner: another user's run id reads as 404, and
    runs recorded before accounts existed (no owner) stay visible to everyone.
    With QUANTLAB_AUTH=off the auth operations answer 404 and every other
    operation behaves exactly as in 0.5.1.

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      summary: Register a user
      # Bootstrap-gated: public while zero users exist, then admin-only.
      security: []
      description: >
        Public only while no users exist; that first user is created with
        `is_admin: true`. Once any user exists, registration requires an
        admin's session cookie and new users are created non-admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Users exist and the caller has no session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Users exist and the caller is not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Username already taken (comparison is case-insensitive)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Username or password outside its declared bounds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Log in, receiving the session cookie
      security: []
      description: >
        On success sets the `quantlab_session` cookie (HttpOnly,
        SameSite=Lax; Secure only when QUANTLAB_AUTH_COOKIE_SECURE is set,
        since local development is plain http). Failed attempts are rate
        limited per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Logged in; the session cookie is set
          headers:
            Set-Cookie:
              description: >
                `quantlab_session=<token>; HttpOnly; SameSite=Lax; Path=/;
                Max-Age=1209600` (`Secure` only behind
                QUANTLAB_AUTH_COOKIE_SECURE)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Username or password outside its declared bounds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many failed attempts from this client IP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Log out, revoking the session
      responses:
        "204":
          description: Session deleted; the cookie is cleared
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: The authenticated user
      responses:
        "200":
          description: The caller's identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: No valid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: quantlab_session
      description: >
        Opaque session token issued by POST /auth/login, stored by the
        backend only as its SHA-256, valid for 14 days since last use.
  schemas:
...
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    AuthCredentials:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 256
    User:
      type: object
      additionalProperties: false
      required: [id, username, is_admin]
      properties:
        id:
          type: integer
        username:
          type: string
        is_admin:
          type: boolean
    UserPublic:
      type: object
      additionalProperties: false
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
...